Auto complete GHC imports etc. via neco-ghc
Unfortunately, this behavior causes auto complete to fail for
definitions in the local module / buffer.

https://github.com/eagletmt/neco-ghc/issues/44

Use `<C-x><C-p>` or `<C-x><C-n>` to manually get that type of
completion.

Also `g:haskellmode_completion_ghc` has to be `0` for me even though
the vim in 2016 article and associated repo show it equal to `1`.

http://www.stephendiehl.com/posts/vim_2016.html